Fix code inspection issues
