+++ src/OSDP.Net/Connections/TcpClientOsdpConnection.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Net.Sockets;
 using System.Threading;
 using System.Threading.Tasks;
